Move remaining aio tests to t_aio.py
