+++ ftd2xx/tests/t_aio.py
@@ -1,6 +1,7 @@
 # This file was originally generated by PyScripter's unitest wizard
 
 import asyncio
+from ftd2xx.ftd2xx import DeviceError
 import unittest
 
 from .. import aio
@@ -30,6 +31,32 @@ class TestAIOFTD2XX(unittest.IsolatedAsyncioTestCase):
                 self.assertIsInstance(e, asyncio.TimeoutError)
                 return
         raise Exception("Could not timeout!")
+
+
+class TestAIOGlobals(unittest.IsolatedAsyncioTestCase):
+    def testaioopen(self):
+        try:
+            device = aio.open()
+            self.assertIsInstance(device, aio.FTD2XX)
+        except AssertionError:
+            raise
+        else:
+            device.close()
+
+    def testaioopenEx(self):
+        dev0 = None
+        try:
+            devices = aio.listDevices()
+            if devices is None:
+                raise DeviceError("Device not found")
+            dev0_id = devices[0]
+            dev0 = aio.openEx(dev0_id)
+            self.assertIsInstance(dev0, aio.FTD2XX)
+        except AssertionError:
+            raise
+        finally:
+            if dev0:
+                dev0.close()
 
 
 # if __name__ == "__main__":
